+++ src/streamer/h264fileparser.cpp
@@ -21,7 +21,11 @@
 
 #include <fstream>
 
+#ifdef _WIN32
+#include <winsock2.h>
+#else
 #include <arpa/inet.h>
+#endif
 
 using namespace std;
